+++ src/main/kotlin/dev/protsenko/securityLinter/kubernetes/quickfix/ReplaceQuickFixes.kt
@@ -9,6 +9,11 @@ class ReplaceValueTo1000QuickFix(@IntentionFamilyName private val message: Strin
 class ReplaceValueToFalseQuickFix(@IntentionFamilyName private val message: String) : ReplaceValueWithEnumQuickFix(message)  {
     override val enumValue: ReplacedValueEnum = ReplacedValueEnum.FALSE
 }
+
 class ReplaceValueToTrueQuickFix(@IntentionFamilyName private val message: String) : ReplaceValueWithEnumQuickFix(message)  {
     override val enumValue: ReplacedValueEnum = ReplacedValueEnum.TRUE
+}
+
+class ReplaceValueToRuntimeDefaultQuickFix(@IntentionFamilyName private val message: String) : ReplaceValueWithEnumQuickFix(message)  {
+    override val enumValue: ReplacedValueEnum = ReplacedValueEnum.RUNTIME_DEFAULT
 }