+++ src/main/kotlin/dev/protsenko/securityLinter/kubernetes/quickfix/ReplacedValueEnum.kt
@@ -1,5 +1,6 @@
 package dev.protsenko.securityLinter.kubernetes.quickfix
 
 enum class ReplacedValueEnum(val value: String) {
-    TRUE("true"), FALSE("false"), ONE_THOUSAND("1000")
+    TRUE("true"), FALSE("false"), ONE_THOUSAND("1000"),
+    RUNTIME_DEFAULT("RuntimeDefault")
 }